moved Event and CustomEvent to browser.window
