+++ lib/Element.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const CustomEvent = require("./CustomEvent");
+const {CustomEvent} = require("./Events");
 const DocumentFragment = require("./DocumentFragment");
 const DOMException = require("domexception");
 const makeAbsolute = require("./makeAbsolute");
