+++ lib/Window.js
@@ -4,6 +4,7 @@ const Element = require("./Element");
 const getLocation = require("./getLocation");
 const Navigator = require("./Navigator");
 const url = require("url");
+const {Event, CustomEvent} = require("./Events");
 const {EventEmitter} = require("events");
 const {performance} = require("perf_hooks");
 
@@ -39,7 +40,8 @@ module.exports = function Window(resp, windowObjects, innerWidth = 760, innerHei
     },
     cancelAnimationFrame: () => {},
     Element,
-    Event
+    Event,
+    CustomEvent
   }, windowObjects);
 
   Object.defineProperty(window, "pageXOffset", {
@@ -98,9 +100,3 @@ module.exports = function Window(resp, windowObjects, innerWidth = 760, innerHei
     window.dispatchEvent("resize");
   }
 };
-
-function Event(type, init = {}) {
-  const event = {...init, type};
-  Object.setPrototypeOf(event, Event.prototype);
-  return event;
-}
