+++ lib/index.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const Compiler = require("./Compiler");
-const CustomEvent = require("./CustomEvent");
 const Document = require("./Document");
 const Fetch = require("./Fetch");
 const IntersectionObserver = require("./IntersectionObserver");
@@ -11,7 +10,6 @@ const Storage = require("./Storage");
 
 module.exports = {
   Compiler,
-  CustomEvent,
   Document,
   Fetch,
   IntersectionObserver,
